+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-﻿## 0.2
+﻿## 0.2.1
+* Get-Extension:  Support for -AllValid (#45)
+* Get-Extension:  Removing Validation Errors when -ErrorAction is ignore (#43)
+* Updating Piecemeal tests: Adding test for Steppable Pipeline (#42)
+* Updating Piecemeal Formatting (making .Extends a list) (re #44)
+* Get-Extension:  Support for SteppablePipelines (#42).  Fixing .Extends bug (#44)
+---
+
+## 0.2
 * Adding formatting for extensions (#40)
 * Updating Piecemeal tests (account for Regex) (#39)
 * Get-Extension:  Regex support for [Cmdlet] attribute (#39)